+++ src/nav-menu/nav-menu.html
@@ -1,57 +1,61 @@
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml">
+
 <head>
-  
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css" />
 </head>
+
 <body>
-  <header v-show="oData.bGuest">      
-    <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow">
+  <!-- <header v-show="oData.bGuest">
+    <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white box-shadow">
+      <img class="logo" src="../assets/images/Barbuuuda.png" alt="">
       <div class="container">
-        <button class="navbar-toggler btn-state-left-menu"
-                type="button"
-                data-toggle="collapse"
-                data-target=".navbar-collapse"
-                aria-label="Toggle navigation"
-                v-on:click="toggle()">
-          <span class="navbar-toggler-icon"></span>
-        </button>
+      <button class="navbar-toggler btn-state-left-menu" type="button" data-toggle="collapse"
+        data-target=".navbar-collapse" aria-label="Toggle navigation" v-on:click="toggle()">
+        <span class="navbar-toggler-icon"></span>
+      </button>
 
-        <div class="navbar-collapse collapse d-sm-inline-flex d-flex header">
-          <!-- Бургер левой панели. -->
-          <div>
-            <div class="block-left-line" v-on:click="onStateLeftPanel()">
-              <div class="left-line"></div>
-              <div class="left-line"></div>
-              <div class="left-line"></div>
-            </div>
+      <div class="navbar-collapse collapse d-sm-inline-flex d-flex header">
+         Бургер левой панели. 
+        <div>
+          <div class="block-left-line" v-on:click="onStateLeftPanel()">
+            <i class="fas fa-bars"></i>
           </div>
+        </div>
 
-          <div>
-            <ul class="navbar-nav flex-grow">
-              <li class="nav-item">
-                <div class="div-link">
-                  <img src="../assets/images/add_task.jpg" class="link"
-                        v-on:click="onRouteCreateTask()" />
-                  <!-- <li><router-link to="" class="nav-link link-add-task"
-                    v-on:click="onRouteCreateTask()">Создать задание</router-link></li> -->
+        <div>
+          <div class="navbar-nav flex-grow">
+            <li class="nav-item">
+              <div class="div-link">
+                 <img src="../assets/images/add_task.jpg" class="link" v-on:click="onRouteCreateTask()" /> 
+                 <li>
+                <router-link to="" class="nav-link link-add-task" v-on:click="onRouteCreateTask()">Создать задание
+                </router-link>
+              </li> 
 
-                    <li>
-                      <button class="nav-link btn btn-link link-add-task" 
-                              v-on:click="onRouteCreateTask()">
-                        Создать задание
-                      </button>
-                    </li>
-                </div>
-              </li>
-            </ul>
+
+                <button class="nav-link btn btn-link link-add-task" v-on:click="onRouteCreateTask()">
+                  <img class="plus" src="../assets/images/plus.png" alt="">
+                  Создать задание
+                </button>
+                <button class="nav-link btn btn-link link-add-task oranged" v-on:click="onCreate()">
+                  <p>Регистрация</p>
+                </button>
+                <button class="nav-link btn btn-link link-add-task oranged" v-on:click="onLogin()">
+                  <p>Вход</p>
+                </button>
+
+              </div>
+            </li>
           </div>
         </div>
       </div>
     </nav>
-  </header>
+  </header>-->
 
-  <!-- Модальное окно ошибки, если нет роли заказчика. -->  
-  <div class="modal fade" id="idNotCustomer" tabindex="-1" role="dialog" aria-labelledby="idNotCustomerTitle" aria-hidden="true">
+  <!-- Модальное окно ошибки, если нет роли заказчика.
+  <div class="modal fade" id="idNotCustomer" tabindex="-1" role="dialog" aria-labelledby="idNotCustomerTitle"
+    aria-hidden="true">
     <div class="modal-dialog modal-dialog-centered" role="document">
       <div class="modal-content">
         <div class="modal-header">
@@ -67,10 +71,12 @@
         </div>
 
         <div class="modal-footer">
-          <button type="button" class="btn-create-task form-control-sm btn-create-task btn-all-tasks" data-dismiss="modal">Хорошо</button>
+          <button type="button" class="btn-create-task form-control-sm btn-create-task btn-all-tasks"
+            data-dismiss="modal">Хорошо</button>
         </div>
       </div>
     </div>
-  </div>
+  </div> -->
 </body>
+
 </html>
